+++ src/main/java/ru/itsjava/service/DiscountCalculationService.java
@@ -3,5 +3,5 @@ package ru.itsjava.service;
 import ru.itsjava.domain.User;
 
 public interface DiscountCalculationService {
-    double calculate(User user, int price);
+    double calculate(User user, double price);
 }
